Add some base classes for client side
It's finally time to start programming! I've got the ideas pretty much done and now I've just got to turn them into code. This is the beginning of that process. There are base classes for all the essential workout data structures. Also there are some display classes. Add list which is one of the main panels in the display code. Next: add support for DivisionSuites to the Session List class, and then move on to what happens when the user double clicks on a List item (begin that exercise/circuit). That will be a switch of screens.