Create transition to ExersizeScreen
When a user double clicks on a list element that is representative of an
Exersize division, that user is sent to the ExersizeScreen--where he may begin
executing sets.

Todo: add support for DivisionSuites. A DivisionSuite is simply a collection of
Exersizes and/or more DivisionSuites. A DivisionSuite belonging to the primary
division is a circuit. Circuits may contain more DivisionSuites. The user will
be brought to a screen similar to the SessionScreen where he may choose from a
list of exersizes or more circuits. Eventually he will bottom out to an
exersize where he is brought to the ExersizeScreen to do his sets for that
exersize, then back to the previous screen where he selects his next exersize.
Circuits take properties of both a list of Exercises and Exercises themselves.
Damn I need to learn how to spell exercise correctly lol. Anyway, they are like
lists because you can select any of the exersizes to do next (but they retain
the order the workout-designer created the list in). DivisionSuites are like
exersizes, because once you start a DivisionSuite you are locked into
completeing that suite. You can either end it early or cancel the DivisionSuite,
but you cannot do part of a division suite, go up a level, finish an exersize,
then come back. The circuit/DivisionSuite is like an exersize for this reason.